improves backend endpoints

+++ src/main/java/ch/zli/m223/punchclock/service/ApplicationUserService.java
@@ -1,7 +1,10 @@
 package ch.zli.m223.punchclock.service;
 
 import ch.zli.m223.punchclock.domain.ApplicationUser;
+import ch.zli.m223.punchclock.domain.UserRole;
+import ch.zli.m223.punchclock.error_handling.ForbiddenException;
 import ch.zli.m223.punchclock.repository.ApplicationUserRepository;
+import org.glassfish.jersey.jaxb.internal.XmlJaxbElementProvider;
 import org.springframework.security.core.userdetails.User;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
@@ -18,11 +21,14 @@ import static java.util.Collections.emptyList;
 public class ApplicationUserService implements UserDetailsService {
     private ApplicationUserRepository applicationUserRepository;
     private BCryptPasswordEncoder bCryptPasswordEncoder;
+    private UserRoleService userRoleService;
 
     public ApplicationUserService(ApplicationUserRepository applicationUserRepository,
-                                  BCryptPasswordEncoder bCryptPasswordEncoder) {
+                                  BCryptPasswordEncoder bCryptPasswordEncoder,
+                                  UserRoleService userRoleService) {
         this.applicationUserRepository = applicationUserRepository;
         this.bCryptPasswordEncoder = bCryptPasswordEncoder;
+        this.userRoleService = userRoleService;
     }
 
     @Override
@@ -35,15 +41,27 @@ public class ApplicationUserService implements UserDetailsService {
     }
 
     public ApplicationUser createUser(ApplicationUser applicationUser){
+        if (userRoleService.getAllUserRoles().size() == 0){
+            UserRole userRole = new UserRole();
+            userRole.setRoleName("Administrator");
+            userRole.setAdmin(true);
+            userRole.setId(Long.valueOf(1));
+            userRoleService.createFirstUserRole(userRole);
+        }
+
         applicationUser.setPassword(bCryptPasswordEncoder.encode(applicationUser.getPassword()));
         return applicationUserRepository.save(applicationUser);
     }
 
-    public List<ApplicationUser> getAllUsers(){
+    public List<ApplicationUser> getAllUsers(ApplicationUser user){
+        if (!user.getRole().getAdmin())
+            throw new ForbiddenException("Only admins allowed");
         return applicationUserRepository.findAll();
     }
 
-    public ApplicationUser getSingleUser(Long id){
+    public ApplicationUser getSingleUser(Long id, ApplicationUser user){
+        if (user.getId() != id && !user.getRole().getAdmin())
+            throw new ForbiddenException("Foreign users not allowed");
         return applicationUserRepository.findById(id).get();
     }
 
@@ -51,12 +69,23 @@ public class ApplicationUserService implements UserDetailsService {
         return applicationUserRepository.findByUsername(username);
     }
 
-    public ApplicationUser editUser(ApplicationUser applicationUser, Long id){
-        applicationUser.setId(id);
-        return applicationUserRepository.save(applicationUser);
+    public ApplicationUser getUserByUsername(String username, ApplicationUser user){
+        ApplicationUser applicationUser = applicationUserRepository.findByUsername(username);
+        if (!user.equals(applicationUser) && !user.getRole().getAdmin())
+            throw new ForbiddenException("Cannot access foreign user");
+        return applicationUser;
     }
 
-    public void deleteUser(Long id){
+    public ApplicationUser editUser(ApplicationUser editedUser, Long id, ApplicationUser user){
+        this.getSingleUser(id, user);
+        editedUser.setId(id);
+        editedUser.setUsername(user.getUsername());
+        return applicationUserRepository.save(editedUser);
+    }
+
+    public void deleteUser(Long id, ApplicationUser user){
+        if (id != user.getId() && !user.getRole().getAdmin())
+            throw new ForbiddenException("Cannot delete foreign users");
         applicationUserRepository.deleteById(id);
     }
 }
